Moving DTOs to DTO package
