+++ data/src/main/java/io/gnosis/data/adapters/MoshiBuilder.kt
@@ -2,6 +2,7 @@ package io.gnosis.data.adapters
 
 import com.squareup.moshi.adapters.PolymorphicJsonAdapterFactory
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import io.gnosis.data.backend.dto.*
 import io.gnosis.data.db.BigDecimalNumberAdapter
 import io.gnosis.data.models.*
 import pm.gnosis.common.adapters.moshi.MoshiBuilderFactory
