+++ data/src/main/java/io/gnosis/data/repositories/TransactionRepository.kt
@@ -1,7 +1,7 @@
 package io.gnosis.data.repositories
 
 import io.gnosis.data.backend.TransactionServiceApi
-import io.gnosis.data.backend.dto.ServiceTokenInfo
+import io.gnosis.data.backend.dto.*
 import io.gnosis.data.models.*
 import io.gnosis.data.repositories.TokenRepository.Companion.ETH_SERVICE_TOKEN_INFO
 import io.gnosis.data.utils.formatBackendDate
@@ -13,7 +13,7 @@ import pm.gnosis.utils.hexToByteArray
 import pm.gnosis.utils.removeHexPrefix
 import java.math.BigInteger
 
-//TODO Remove these fake tokens, when we have ContractInfo from the backend
+// TODO Remove these fake tokens, when we have ContractInfo from the backend
 private val defaultErc20Address = "0xc778417e063141139fce010982780140aa0cd5ab".asEthereumAddress()!!
 private val defaultErc721Address = "0xB3775fB83F7D12A36E0475aBdD1FCA35c091efBe".asEthereumAddress()!!
 val FAKE_ERC20_TOKEN_INFO = ServiceTokenInfo(defaultErc20Address, 18, "WETH", "Wrapped Ether", "local::ethereum")
@@ -22,7 +22,6 @@ val FAKE_ERC721_TOKEN_INFO = ServiceTokenInfo(defaultErc721Address, 18, "DRK", "
 class TransactionRepository(
     private val transactionServiceApi: TransactionServiceApi
 ) {
-
 
     suspend fun getTransactions(safeAddress: Solidity.Address): Page<Transaction> =
         transactionServiceApi.loadTransactions(safeAddress.asEthereumAddressChecksumString())
@@ -65,9 +64,9 @@ class TransactionRepository(
         transactionDto.data.hexStringNullOrEmpty() && transactionDto.operation == Operation.CALL
 
     private fun isSettingsChange(transactionDto: MultisigTransactionDto): Boolean =
-        transactionDto.to == transactionDto.safe
-                && transactionDto.operation == Operation.CALL
-                && SafeRepository.isSettingsMethod(transactionDto.dataDecoded?.method)
+        transactionDto.to == transactionDto.safe &&
+                transactionDto.operation == Operation.CALL &&
+                SafeRepository.isSettingsMethod(transactionDto.dataDecoded?.method)
 
     private fun transfer(transferDto: TransferDto): Transaction.Transfer {
         return Transaction.Transfer(
@@ -94,7 +93,7 @@ class TransactionRepository(
             ETH_SERVICE_TOKEN_INFO
         )
 
-    //when contractInfo is available have when for ETH, ERC20 and ERC721
+    // when contractInfo is available have when for ETH, ERC20 and ERC721
     private fun transferEth(transaction: MultisigTransactionDto): Transaction.Transfer =
         Transaction.Transfer(
             transaction.to,
@@ -108,7 +107,7 @@ class TransactionRepository(
         val to = transaction.dataDecoded?.parameters?.getValueByName("to")?.asEthereumAddress() ?: Solidity.Address(BigInteger.ZERO)
         val value = transaction.dataDecoded?.parameters?.getValueByName("value")?.decimalAsBigInteger() ?: BigInteger.ZERO
 
-        //Only available with transferFrom
+        // Only available with transferFrom
         val from = transaction.dataDecoded?.parameters?.getValueByName("from")?.asEthereumAddress()
 
         return Transaction.Transfer(
@@ -173,7 +172,6 @@ class TransactionRepository(
             null,
             BigInteger.ZERO
         )
-
 }
 
 fun List<ParamsDto>?.getValueByName(name: String): String? {
