+++ data/src/test/java/io/gnosis/data/repositories/TransactionRepositoryTest.kt
@@ -1,6 +1,7 @@
 package io.gnosis.data.repositories
 
 import io.gnosis.data.backend.TransactionServiceApi
+import io.gnosis.data.backend.dto.*
 import io.gnosis.data.models.*
 import io.gnosis.data.repositories.TokenRepository.Companion.ETH_SERVICE_TOKEN_INFO
 import io.mockk.coEvery
@@ -49,7 +50,7 @@ class TransactionRepositoryTest {
 
         with(actual.results[0] as Transaction.Custom) {
             assertEquals(transactionDto.module, address)
-            assertEquals( 0L, dataSize)
+            assertEquals(0L, dataSize)
             assertEquals(transactionDto.created, date)
             assertEquals(transactionDto.value, value)
         }
@@ -67,7 +68,7 @@ class TransactionRepositoryTest {
         assertEquals(1, actual.results.size)
         with(actual.results[0] as Transaction.Custom) {
             assertEquals(transactionDto.to, address)
-            assertEquals( 0L, dataSize)
+            assertEquals(0L, dataSize)
         }
     }
 
@@ -171,7 +172,6 @@ class TransactionRepositoryTest {
             //  TODO: check for right transfer type
             assertEquals(FAKE_ERC721_TOKEN_INFO, tokenInfo)
         }
-
     }
 
     @Test
